move Container.php to Auth/Container.php
